Refactor Program to allow overriding the IConsoleManager
This will enable tests to run without being tied to the ConsoleManager implementation, and hence to the Console concrete dependency (mostly).  It opens the doors to allowing us to drive the REPL in test code, and be able to read the console output back out.

This also works around the check for Console.IsOutputRedirected, as the state of that property is determined by the test execution engine.  When launched under a debugger, IsOutputRedirected is false (which is good), but when launched not under a debugger IsOutputRedirected is true - this previously made it impossible for tests to enter the REPL execution loop from this entrypoint.