Add an integration test with the Swagger endpoint disabled
This adds a flag to the test fixture to toggle whether the Swagger
endpoint is active, which allows tests to be written with it on or off.

Still to do: find a way to toggle it on each test run.  Ideally, we
would have a single Theory test which would take as inputs whether the
Swagger endpoint is active, and the corresponding expected output.
Currently the tests have to be separate because the server is started
during Fixture initialization, which is before the test method receives
control.